// src/enums/ContextActions.ts
export enum ContextActions {
	setCurrentUser = 'SET_CURRENT_USER',
	setComments = 'SET_COMMENTS',
	addComment = 'ADD_COMMENT',
	updateComment = 'UPDATE_COMMENT',
	deleteComment = 'DELETE_COMMENT',
	addReply = 'ADD_REPLY',
	updateReply = 'UPDATE_REPLY',
	deleteReply = 'DELETE_REPLY',
	updateCommentScore = 'UPDATE_COMMENT_SCORE',
	updateReplyScore = 'UPDATE_REPLY_SCORE',
}